renegade_wallet_client: websocket: lazy-init single task history subscription (#53)
In this PR, we transition the `RenegadeWebsocketClient` to using the task history websocket topic rather than the task status one. We lazily instantiate a single subscription to the wallet's task history topic the first time that a task is watched. This is more reliable than subscribing to per-task status topics, and effectively removes the need for ever fetching task history from the historical state engine.

### Testing
- [x] Test subscribing to multiple successive tasks on the same client via the synthetic tester