renegade_wallet_client: websocket: task_waiter: configurable timeout (#54)
In this PR, we allow the timeout used by the `TaskWaiter` to be configurable via a `TaskWaiterBuilder`. We bump the default timeout to one minute and use a timeout of 2 minutes for the withdrawal task created by the `withdraw` action, as the waiter for that task may be awaited while any preceding fee payment tasks are still executing.

### Testing

- [x] Test via local synthetic tester