renegade_wallet_client: actions: withdraw: expect fee payment in withdrawal endpoint (#57)
In this PR, we update the withdrawal action to expect fee payments to be enqueued by the relayer API, rather than enqueueing fees ourselves. Instead, we replicate the effects of fee payments on the local view of the wallet before applying the withdrawal changes, to ensure the correct post-wallet is signed.

### Testing

- [x] Test against local relayer w/ mock wallets containing pre-allocated fees